+++ src/super/i-block/i-block.ts
@@ -50,7 +50,6 @@ import {
 	RawWatchHandler,
 
 	Hook,
-	ActivationStatus,
 
 	ComponentInterface,
 	UnsafeGetter,
@@ -2234,11 +2233,6 @@ export default abstract class iBlock extends ComponentInterface {
 	 */
 	deactivate(): void {
 		deactivate(this);
-	}
-
-	/** @override */
-	emitActivation(status: ActivationStatus): Promise<void> {
-		return this.r.emitActivation(status);
 	}
 
 	/**